Changed the database columns to snake_case and id type to long

+++ src/main/java/net/therap/hibernet/service/UserService.java
@@ -26,11 +26,11 @@ public class UserService {
         return userDao.getAll(query);
     }
 
-    public User getUser(int id) {
+    public User getUser(long id) {
         return entityManager.find(User.class, id);
     }
 
-    public void addUser(int id, String name) {
+    public void addUser(long id, String name) {
         User user = new User();
         user.setId(id);
         user.setName(name);
@@ -38,13 +38,13 @@ public class UserService {
         userDao.add(user);
     }
 
-    public void updateUser(int id, String newName) {
+    public void updateUser(long id, String newName) {
         User user = entityManager.find(User.class, id);
 
         userDao.update(user, newName);
     }
 
-    public void deleteUser(int id) {
+    public void deleteUser(long id) {
         User user = entityManager.find(User.class, id);
 
         userDao.delete(user);
